pyproject.toml
[project]
name = "exp-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "scikit-learn>=1.3.0",
]
